+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## Unreleased
+
+- Added suggestion to add `loginPath` to `config/craft/general.php`.
+
 ## 1.0.0 - 2026-05-06
 
 - Initial release with support for Craft CMS 6.0.0-alpha.1